refactor(cloud-toolkit): mv action-runner  to dialogs folder (#988)
